LPS-105380 Move Configuraton class to package *.configuration
