+++ modules/apps/static/portal-bundle-blacklist/portal-bundle-blacklist-impl/src/main/java/com/liferay/portal/bundle/blacklist/internal/BundleBlacklist.java
@@ -15,6 +15,7 @@
 package com.liferay.portal.bundle.blacklist.internal;
 
 import com.liferay.osgi.util.BundleUtil;
+import com.liferay.portal.bundle.blacklist.internal.configuration.BundleBlacklistConfiguration;
 import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
@@ -54,7 +55,7 @@ import org.osgi.service.component.annotations.Reference;
  * @author Matthew Tambara
  */
 @Component(
-	configurationPid = "com.liferay.portal.bundle.blacklist.internal.BundleBlacklistConfiguration",
+	configurationPid = "com.liferay.portal.bundle.blacklist.internal.configuration.BundleBlacklistConfiguration",
 	immediate = true, service = BundleBlacklist.class
 )
 public class BundleBlacklist {
