+++ modules/apps/static/portal-bundle-blacklist/portal-bundle-blacklist-test/src/testIntegration/java/com/liferay/portal/bundle/blacklist/test/BundleBlacklistTest.java
@@ -399,7 +399,7 @@ public class BundleBlacklistTest {
 		"com.liferay.portal.bundle.blacklist.internal.BundleBlacklist";
 
 	private static final String _CONFIG_NAME =
-		"com.liferay.portal.bundle.blacklist.internal." +
+		"com.liferay.portal.bundle.blacklist.internal.configuration." +
 			"BundleBlacklistConfiguration";
 
 	private static final String _LPKG_NAME = "Bundle Blacklist Test";
