+++ modules/apps/static/portal-component-blacklist/portal-component-blacklist-impl/src/main/java/com/liferay/portal/component/blacklist/internal/configuration/ComponentBlacklistConfiguration.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.component.blacklist.internal;
+package com.liferay.portal.component.blacklist.internal.configuration;
 
 import aQute.bnd.annotation.metatype.Meta;
 
@@ -23,7 +23,7 @@ import com.liferay.portal.configuration.metatype.annotations.ExtendedObjectClass
  */
 @ExtendedObjectClassDefinition(category = "module-container")
 @Meta.OCD(
-	id = "com.liferay.portal.component.blacklist.internal.ComponentBlacklistConfiguration",
+	id = "com.liferay.portal.component.blacklist.internal.configuration.ComponentBlacklistConfiguration",
 	localization = "content/Language",
 	name = "component-blacklist-configuration-name"
 )
